chore: remove FindOrganisation component tied to legacy search path
Component was only used when `FilteredSearch` was disabled. Reused inputs and supporting files moved to `/components/suggesters`.
