+++ front-end-components/src/components/suggesters/index.tsx
@@ -1,0 +1,5 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from "src/components/suggesters/types";
+export * from "src/components/suggesters/la-input";
+export * from "src/components/suggesters/school-input";
+export * from "src/components/suggesters/trust-input";
