--- front-end-components/src/views/find-organisation/index.tsx
@@ -1,3 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export * from "src/views/find-organisation/types";
-export * from "src/views/find-organisation/view";
