+++ front-end-components/src/views/index.tsx
@@ -5,6 +5,5 @@ export * from "src/views/compare-your-census";
 export * from "src/views/historic-data";
 export * from "src/views/historic-data-2";
 export * from "src/views/deployment-plan";
-export * from "src/views/find-organisation";
 export * from "src/views/la-national-rank";
 export * from "src/views/benchmark-data-high-needs";
